+++ spec/install/security_policy_spec.rb
@@ -14,37 +14,37 @@ describe "policies with unsigned gems" do
     G
   end
 
-  it "works after you try to deploy without a lock" do
+  it "will work after you try to deploy without a lock" do
     bundle "install --deployment"
     bundle :install, :exitstatus => true
     expect(exitstatus).to eq(0)
     should_be_installed "rack 1.0", "signed_gem 1.0"
   end
 
-  it "fails when given invalid security policy" do
+  it "will fail when given invalid security policy" do
     bundle "install --trust-policy=InvalidPolicyName"
     expect(out).to include("Rubygems doesn't know about trust policy")
   end
 
-  it "fails with High Security setting due to presence of unsigned gem" do
+  it "will fail with High Security setting due to presence of unsigned gem" do
     bundle "install --trust-policy=HighSecurity"
     expect(out).to include("security policy didn't allow")
   end
 
   # This spec will fail on Rubygems 2 rc1 due to a bug in policy.rb. the bug is fixed in rc3.
-  it "fails with Medium Security setting due to presence of unsigned gem", :unless => ENV['RGV'] == "v2.0.0.rc.1" do
+  it "will fail with Medium Security setting due to presence of unsigned gem", :unless => ENV['RGV'] == "v2.0.0.rc.1" do
     bundle "install --trust-policy=MediumSecurity"
     expect(out).to include("security policy didn't allow")
   end
 
-  it "succeeds with no policy" do
+  it "will succeed with no policy" do
     bundle "install", :exitstatus => true
     expect(exitstatus).to eq(0)
   end
 
 end
 
-describe "policies with signed gems, no CA" do
+describe "policies with signed gems and no CA" do
   before do
     build_security_repo
     gemfile <<-G
@@ -53,26 +53,25 @@ describe "policies with signed gems, no CA" do
     G
   end
 
-  it "fails with High Security setting, gem is self-signed" do
+  it "will fail with High Security setting, gem is self-signed" do
     bundle "install --trust-policy=HighSecurity"
     expect(out).to include("security policy didn't allow")
   end
 
-  it "fails with Medium Security setting, gem is self-signed" do
+  it "will fail with Medium Security setting, gem is self-signed" do
     bundle "install --trust-policy=MediumSecurity"
     expect(out).to include("security policy didn't allow")
   end
 
-  it "succeeds with Low Security setting, low security accepts self signed gem" do
+  it "will succeed with Low Security setting, low security accepts self signed gem" do
     bundle "install --trust-policy=LowSecurity", :exitstatus => true
     expect(exitstatus).to eq(0)
     should_be_installed "signed_gem 1.0"
   end
 
-  it "succeeds with no policy" do
+  it "will succeed with no policy" do
     bundle "install", :exitstatus => true
     expect(exitstatus).to eq(0)
     should_be_installed "signed_gem 1.0"
   end
-
 end