+++ spec/other/check_spec.rb
@@ -23,7 +23,7 @@ describe "bundle check" do
     expect(out).to eq("The Gemfile's dependencies are satisfied")
   end
 
-  it "creates a Gemfile.lock by default if one did not exist" do
+  it "creates a Gemfile.lock by default if one does not exist" do
     install_gemfile <<-G
       source "file://#{gem_repo1}"
       gem "rails"