+++ spec/quality_spec.rb
@@ -5,6 +5,18 @@ if defined?(Encoding) && Encoding.default_external != "UTF-8"
 end
 
 describe "The library itself" do
+  def check_for_spec_defs_with_single_quotes(filename)
+    failing_lines = []
+
+    File.readlines(filename).each_with_index do |line,number|
+      failing_lines << number + 1 if line =~ /^ *(describe|it|context) {1}'{1}/
+    end
+
+    unless failing_lines.empty?
+      "#{filename} uses inconsistent single quotes on lines #{failing_lines.join(', ')}"
+    end
+  end
+
   def check_for_tab_characters(filename)
     failing_lines = []
     File.readlines(filename).each_with_index do |line,number|
@@ -46,6 +58,18 @@ describe "The library itself" do
         next if filename =~ exempt
         error_messages << check_for_tab_characters(filename)
         error_messages << check_for_extra_spaces(filename)
+      end
+    end
+    expect(error_messages.compact).to be_well_formed
+  end
+
+  it "uses double-quotes consistently in specs" do
+    included = /spec/
+    error_messages = []
+    Dir.chdir(File.expand_path("../", __FILE__)) do
+      `git ls-files`.split("\n").each do |filename|
+        next unless filename =~ included
+        error_messages << check_for_spec_defs_with_single_quotes(filename)
       end
     end
     expect(error_messages.compact).to be_well_formed